Save discount value in item if any

+++ src/WWS/Inventory/inventory/src/main/kotlin/hu/bme/aut/inventory/controller/review/response/ReviewResponse.kt
@@ -19,7 +19,7 @@ data class ReviewResponse(
                 reviewerName = reviewerName,
                 reviewerId = reviewerId,
                 summary = summary,
-                rating = rating,
+                rating = rating.toInt(),
                 created = created.toString()
             )
         }
